+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [fj-doc-maven-plugin] new param 'addVerifyPlugin' (default 'true') of 'add' plugin
+
 ## [8.7.0] - 2024-08-25
 
 ### Added